% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_event.R
\name{mp_event}
\alias{mp_event}
\title{Create a Measurement Protocol Event}
\usage{
mp_event(name, params = NULL, items = NULL)
}
\arguments{
\item{name}{The event name to send in}

\item{params}{Optional event parameters sent in as a named list}

\item{items}{Optional items created via \link{mp_event_item}}
}
\value{
An \code{mp_event} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This creates an event to send via \link{mp_send}
}
\examples{

mp_event("custom_event")
mp_event("custom_event", params = list(my_param = "SUPER"))
}
\seealso{
Other Measurement Protocol functions: 
\code{\link{mp_cid}()},
\code{\link{mp_event_item}()},
\code{\link{mp_send}()}
}
\concept{Measurement Protocol functions}
