% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_trackme.R
\name{mp_opt_in}
\alias{mp_opt_in}
\title{Tracking opt-in for this package}
\usage{
mp_opt_in()
}
\value{
No return value, called for side effects
}
\description{
This is the opt-in function for this package, using \link{mp_trackme}
}
