% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_trackme.R
\name{mp_trackme_startup}
\alias{mp_trackme_startup}
\title{Package tracking opt-in startup message}
\usage{
mp_trackme_startup(package)
}
\arguments{
\item{package}{The package to set-up tracking for}
}
\description{
Place in .onAttach to have a message appear on first startup
}
